server/service: log GRPC operations
When we try to perform a signing, weight, or threshold operation over GRPC, log
it. This includes storing any relevant HSM device logs in the database,
associating them with the log entry for this action, and attaching the
application-level data associated with this log entry. It also requires
calculating the hash for this log entry, which is based on the contents of this
log entry and the hash of the previous log entry. If this is the first log
entry, add a "blank" log entry with a hash value derived from random
data.